v2.0.0

+++ application/handler/caddy/vhost.go
@@ -1,19 +1,19 @@
 /*
+   Nging is a toolbox for webmasters
+   Copyright (C) 2018-present  Wenhui Shen <swh@admpub.com>
 
-   Copyright 2016 Wenhui Shen <www.webx.top>
+   This program is free software: you can redistribute it and/or modify
+   it under the terms of the GNU Affero General Public License as published
+   by the Free Software Foundation, either version 3 of the License, or
+   (at your option) any later version.
 
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU Affero General Public License for more details.
 
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-
+   You should have received a copy of the GNU Affero General Public License
+   along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 package caddy
 
@@ -24,45 +24,71 @@ import (
 	"net/url"
 	"os"
 	"path/filepath"
-
 	"strings"
 
+	"github.com/admpub/log"
 	"github.com/admpub/nging/application/dbschema"
 	"github.com/admpub/nging/application/handler"
 	"github.com/admpub/nging/application/library/config"
 	"github.com/admpub/nging/application/library/filemanager"
-	"github.com/admpub/nging/application/library/modal"
 	"github.com/admpub/nging/application/library/notice"
 	"github.com/admpub/nging/application/model"
+	"github.com/webx-top/com"
 	"github.com/webx-top/db"
 	"github.com/webx-top/echo"
-	"github.com/webx-top/echo/handler/mvc/events"
+	"github.com/webx-top/echo/middleware/tplfunc"
 )
-
-func init() {
-	handler.RegisterToGroup(`/manage`, func(g *echo.Group) {
-		g.Route(`GET`, ``, VhostIndex)
-		g.Route(`GET,POST`, `/vhost_add`, VhostAdd)
-		g.Route(`GET,POST`, `/vhost_edit`, VhostEdit)
-		g.Route(`GET,POST`, `/vhost_delete`, VhostDelete)
-		g.Route(`GET,POST`, `/vhost_file`, VhostFile)
-		g.Route(`GET`, `/vhost_build`, Vhostbuild)
-		g.Route(`GET`, `/clear_cache`, ClearCache)
-	})
-}
 
 func VhostIndex(ctx echo.Context) error {
 	m := model.NewVhost(ctx)
-	page, size, totalRows, p := handler.PagingWithPagination(ctx)
-	cnt, err := m.List(nil, nil, page, size)
-	ret := handler.Err(ctx, err)
-	if totalRows <= 0 {
-		totalRows = int(cnt())
-		p.SetRows(totalRows)
+	groupID := ctx.Formx(`groupId`).Uint()
+	cond := db.Cond{}
+	if groupID > 0 {
+		cond[`group_id`] = groupID
 	}
-	ctx.Set(`pagination`, p)
-	ctx.Set(`listData`, m.Objects())
-	return ctx.Render(`manage/index`, ret)
+	_, err := handler.PagingWithLister(ctx, handler.NewLister(m, nil, func(r db.Result) db.Result {
+		return r.OrderBy(`-id`)
+	}, cond))
+	ret := handler.Err(ctx, err)
+	rows := m.Objects()
+	gIds := []uint{}
+	rowAndGroup := make([]*model.VhostAndGroup, len(rows))
+	for k, u := range rows {
+		rowAndGroup[k] = &model.VhostAndGroup{
+			Vhost: u,
+		}
+		if u.GroupId < 1 {
+			continue
+		}
+		if !com.InUintSlice(u.GroupId, gIds) {
+			gIds = append(gIds, u.GroupId)
+		}
+	}
+
+	mg := &dbschema.VhostGroup{}
+	var groupList []*dbschema.VhostGroup
+	if len(gIds) > 0 {
+		_, err = mg.List(&groupList, nil, 1, 1000, db.Cond{`id IN`: gIds})
+		if err != nil {
+			if ret == nil {
+				ret = err
+			}
+		} else {
+			for k, v := range rowAndGroup {
+				for _, g := range groupList {
+					if g.Id == v.GroupId {
+						rowAndGroup[k].Group = g
+						break
+					}
+				}
+			}
+		}
+	}
+	ctx.Set(`listData`, rowAndGroup)
+	mg.ListByOffset(&groupList, nil, 0, -1)
+	ctx.Set(`groupList`, groupList)
+	ctx.Set(`groupId`, groupID)
+	return ctx.Render(`caddy/vhost`, ret)
 }
 
 func Vhostbuild(ctx echo.Context) error {
@@ -73,14 +99,18 @@ func Vhostbuild(ctx echo.Context) error {
 				return err
 			}
 			if info.IsDir() {
-				return filepath.SkipDir
+				return nil
 			}
+			if !strings.HasSuffix(path, `.conf`) {
+				return nil
+			}
+			log.Info(`Delete the Caddy configuration file: `, path)
 			return os.Remove(path)
 		})
 	}
 	if err != nil {
 		handler.SendFail(ctx, err.Error())
-		return ctx.Redirect(`/manage`)
+		return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
 	}
 	m := model.NewVhost(ctx)
 	n := 100
@@ -90,7 +120,7 @@ func Vhostbuild(ctx echo.Context) error {
 			_, err = m.ListByOffset(nil, nil, i, n, `disabled`, `N`)
 			if err != nil {
 				handler.SendFail(ctx, err.Error())
-				return ctx.Redirect(`/manage`)
+				return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
 			}
 		}
 		for _, m := range m.Objects() {
@@ -102,25 +132,27 @@ func Vhostbuild(ctx echo.Context) error {
 			}
 			if err != nil {
 				handler.SendFail(ctx, err.Error())
-				return ctx.Redirect(`/manage`)
+				return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
 			}
 		}
 	}
-	err = config.DefaultCLIConfig.CaddyRestart()
+	err = config.DefaultCLIConfig.CaddyReload()
 	if err != nil {
 		ctx.Logger().Error(err)
 	}
 	handler.SendOk(ctx, ctx.T(`操作成功`))
-	return ctx.Redirect(`/manage`)
+	return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
 }
 
 func VhostAdd(ctx echo.Context) error {
 	var err error
+	m := model.NewVhost(ctx)
 	if ctx.IsPost() {
-		m := model.NewVhost(ctx)
 		m.Domain = ctx.Form(`domain`)
 		m.Disabled = ctx.Form(`disabled`)
 		m.Root = ctx.Form(`root`)
+		m.Name = ctx.Form(`name`)
+		m.GroupId = ctx.Formx(`groupId`).Uint()
 		var b []byte
 		b, err = json.Marshal(ctx.Forms())
 		switch {
@@ -132,22 +164,40 @@ func VhostAdd(ctx echo.Context) error {
 			}
 			fallthrough
 		case 0 == 1:
-			err = saveVhostData(ctx, m.Vhost, ctx.Forms())
+			err = saveVhostData(ctx, m.Vhost, ctx.Forms(), false)
 		}
 		if err == nil {
 			handler.SendOk(ctx, ctx.T(`操作成功`))
-			return ctx.Redirect(`/manage`)
+			return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
 		}
-
-		ctx.SetFunc(`Val`, func(name, defaultValue string) string {
-			return ctx.Form(name)
-		})
 	} else {
-		ctx.SetFunc(`Val`, func(name, defaultValue string) string {
-			return defaultValue
-		})
+		id := ctx.Formx(`copyId`).Uint()
+		if id > 0 {
+			err = m.Get(nil, db.Cond{`id`: id})
+			if err == nil {
+				var formData url.Values
+				if e := json.Unmarshal([]byte(m.Setting), &formData); e == nil {
+					for key, values := range formData {
+						for k, v := range values {
+							if k == 0 {
+								ctx.Request().Form().Set(key, v)
+								continue
+							}
+							ctx.Request().Form().Add(key, v)
+						}
+					}
+				}
+				ctx.Request().Form().Set(`id`, `0`)
+			}
+		}
 	}
-	return ctx.Render(`manage/vhost_edit`, err)
+	ctx.SetFunc(`Val`, func(name, defaultValue string) string {
+		return ctx.Form(name, defaultValue)
+	})
+	g := &dbschema.VhostGroup{}
+	g.ListByOffset(nil, nil, 0, -1)
+	ctx.Set(`groupList`, g.Objects())
+	return ctx.Render(`caddy/vhost_edit`, err)
 }
 
 func getSaveDir() (saveFile string, err error) {
@@ -167,15 +217,17 @@ func getSaveDir() (saveFile string, err error) {
 func saveVhostConf(ctx echo.Context, saveFile string, values url.Values) error {
 	SetCaddyfileFunc(ctx, values)
 	ctx.Set(`values`, values)
-	b, err := ctx.Fetch(`manage/caddyfile`, nil)
+	b, err := ctx.Fetch(`caddy/caddyfile`, nil)
 	if err != nil {
 		return err
 	}
+	b = com.CleanSpaceLine(b)
+	log.Info(`Generate a Caddy configuration file: `, saveFile)
 	err = ioutil.WriteFile(saveFile, b, os.ModePerm)
 	return err
 }
 
-func saveVhostData(ctx echo.Context, m *dbschema.Vhost, values url.Values) (err error) {
+func saveVhostData(ctx echo.Context, m *dbschema.Vhost, values url.Values, restart bool) (err error) {
 	var saveFile string
 	saveFile, err = getSaveDir()
 	if err != nil {
@@ -189,9 +241,9 @@ func saveVhostData(ctx echo.Context, m *dbschema.Vhost, values url.Values) (err 
 		}
 	} else {
 		err = saveVhostConf(ctx, saveFile, values)
-		if len(ctx.Form(`restart`)) > 0 {
-			err = config.DefaultCLIConfig.CaddyRestart()
-		}
+	}
+	if err == nil && restart {
+		err = config.DefaultCLIConfig.CaddyReload()
 	}
 	return
 }
@@ -200,34 +252,40 @@ func VhostDelete(ctx echo.Context) error {
 	id := ctx.Formx(`id`).Uint()
 	if id < 1 {
 		handler.SendFail(ctx, ctx.T(`id无效`))
-		return ctx.Redirect(`/manage`)
+		return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
 	}
 	m := model.NewVhost(ctx)
 	err := m.Delete(nil, db.Cond{`id`: id})
 	if err != nil {
 		handler.SendFail(ctx, err.Error())
 	} else {
-		var saveFile string
-		saveFile, err = filepath.Abs(config.DefaultConfig.Sys.VhostsfileDir)
+		err = DeleteCaddyfileByID(id)
 		if err == nil {
-			saveFile = filepath.Join(saveFile, fmt.Sprint(id)+`.conf`)
-			err = os.Remove(saveFile)
-			if os.IsNotExist(err) {
-				err = nil
-			}
-			if err == nil {
-				handler.SendOk(ctx, ctx.T(`操作成功`))
-			}
+			handler.SendOk(ctx, ctx.T(`操作成功`))
 		}
 	}
-	return ctx.Redirect(`/manage`)
+	return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
+}
+
+func DeleteCaddyfileByID(id uint) error {
+	saveFile, err := filepath.Abs(config.DefaultConfig.Sys.VhostsfileDir)
+	if err == nil {
+		saveFile = filepath.Join(saveFile, fmt.Sprint(id)+`.conf`)
+		err = os.Remove(saveFile)
+		if err == nil {
+			err = config.DefaultCLIConfig.CaddyReload()
+		} else if os.IsNotExist(err) {
+			err = nil
+		}
+	}
+	return err
 }
 
 func VhostEdit(ctx echo.Context) error {
 	id := ctx.Formx(`id`).Uint()
 	if id < 1 {
 		handler.SendFail(ctx, ctx.T(`id无效`))
-		return ctx.Redirect(`/manage`)
+		return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
 	}
 
 	var err error
@@ -235,12 +293,14 @@ func VhostEdit(ctx echo.Context) error {
 	err = m.Get(nil, db.Cond{`id`: id})
 	if err != nil {
 		handler.SendFail(ctx, err.Error())
-		return ctx.Redirect(`/manage`)
+		return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
 	}
 	if ctx.IsPost() {
 		m.Domain = ctx.Form(`domain`)
 		m.Disabled = ctx.Form(`disabled`)
 		m.Root = ctx.Form(`root`)
+		m.Name = ctx.Form(`name`)
+		m.GroupId = ctx.Formx(`groupId`).Uint()
 		var b []byte
 		b, err = json.Marshal(ctx.Forms())
 		switch {
@@ -252,45 +312,72 @@ func VhostEdit(ctx echo.Context) error {
 			}
 			fallthrough
 		case 0 == 1:
-			err = saveVhostData(ctx, m.Vhost, ctx.Forms())
+			err = saveVhostData(ctx, m.Vhost, ctx.Forms(), len(ctx.Form(`restart`)) > 0)
 		}
 		if err == nil {
 			handler.SendOk(ctx, ctx.T(`操作成功`))
-			return ctx.Redirect(`/manage`)
+			return ctx.Redirect(handler.URLFor(`/caddy/vhost`))
+		}
+	} else if ctx.IsAjax() {
+		disabled := ctx.Query(`disabled`)
+		if len(disabled) > 0 {
+			m.Disabled = disabled
+			data := ctx.Data()
+			err = m.SetField(nil, `disabled`, disabled, db.Cond{`id`: id})
+			if err != nil {
+				data.SetError(err)
+				return ctx.JSON(data)
+			}
+			if m.Disabled == `Y` {
+				err = DeleteCaddyfileByID(id)
+			} else {
+				var formData url.Values
+				err = json.Unmarshal([]byte(m.Setting), &formData)
+				if err == nil {
+					for name, fun := range tplfunc.TplFuncMap {
+						ctx.SetFunc(name, fun)
+					}
+					err = saveVhostData(ctx, m.Vhost, formData, true)
+				}
+			}
+			if err != nil {
+				data.SetError(err)
+				return ctx.JSON(data)
+			}
+			data.SetInfo(ctx.T(`操作成功`))
+			return ctx.JSON(data)
 		}
 	} else {
 		var formData url.Values
 		if e := json.Unmarshal([]byte(m.Setting), &formData); e == nil {
 			for key, values := range formData {
-				for _, v := range values {
+				for k, v := range values {
+					if k == 0 {
+						ctx.Request().Form().Set(key, v)
+						continue
+					}
 					ctx.Request().Form().Add(key, v)
 				}
 			}
 		}
 	}
 	ctx.SetFunc(`Val`, func(name, defaultValue string) string {
-		return ctx.Form(name)
+		return ctx.Form(name, defaultValue)
 	})
-	ctx.Set(`activeURL`, `/manage`)
-	return ctx.Render(`manage/vhost_edit`, err)
-}
-
-func ClearCache(ctx echo.Context) error {
-	if err := modal.Clear(); err != nil {
-		return err
-	}
-	notice.Clear()
-	events.Event(`clearCache`, func(_ bool) {})
-	return ctx.String(ctx.T(`已经清理完毕`))
+	ctx.Set(`activeURL`, `/caddy/vhost`)
+	g := &dbschema.VhostGroup{}
+	g.ListByOffset(nil, nil, 0, -1)
+	ctx.Set(`groupList`, g.Objects())
+	return ctx.Render(`caddy/vhost_edit`, err)
 }
 
 func VhostFile(ctx echo.Context) error {
 	var err error
-	vhostId := ctx.Formx(`id`).Uint()
+	id := ctx.Formx(`id`).Uint()
 	filePath := ctx.Form(`path`)
 	do := ctx.Form(`do`)
 	m := model.NewVhost(ctx)
-	err = m.Get(nil, db.Cond{`id`: vhostId})
+	err = m.Get(nil, db.Cond{`id`: id})
 	mgr := filemanager.New(m.Root, config.DefaultConfig.Sys.EditableFileMaxBytes, ctx)
 	absPath := m.Root
 	if err == nil && len(m.Root) > 0 {
@@ -341,7 +428,7 @@ func VhostFile(ctx echo.Context) error {
 					notice.OpenMessage(user.Username, `upload`)
 					notice.Send(user.Username, notice.NewMessageWithValue(`upload`, ctx.T(`文件上传出错`), err.Error()))
 				}
-				return err
+				return ctx.JSON(echo.H{`error`: err.Error()}, 500)
 			}
 			return ctx.String(`OK`)
 		default:
@@ -357,14 +444,25 @@ func VhostFile(ctx echo.Context) error {
 	if filePath == `.` {
 		filePath = ``
 	}
+	pathSlice := strings.Split(strings.Trim(filePath, echo.FilePathSeparator), echo.FilePathSeparator)
+	pathLinks := make(echo.KVList, len(pathSlice))
+	encodedSep := filemanager.EncodedSepa
+	urlPrefix := fmt.Sprintf(`/caddy/vhost_file?id=%d&path=`, id) + encodedSep
+	for k, v := range pathSlice {
+		urlPrefix += com.URLEncode(v)
+		pathLinks[k] = &echo.KV{K: v, V: urlPrefix}
+		urlPrefix += encodedSep
+	}
+	ctx.Set(`pathLinks`, pathLinks)
+	ctx.Set(`rootPath`, strings.TrimSuffix(m.Root, echo.FilePathSeparator))
 	ctx.Set(`path`, filePath)
 	ctx.Set(`absPath`, absPath)
 	ctx.SetFunc(`Editable`, func(fileName string) bool {
 		_, ok := Editable(fileName)
 		return ok
 	})
-	ctx.Set(`activeURL`, `/manage`)
-	return ctx.Render(`manage/file`, err)
+	ctx.Set(`activeURL`, `/caddy/vhost`)
+	return ctx.Render(`caddy/file`, err)
 }
 
 func Editable(fileName string) (string, bool) {
